Code Review: defect 103 "ipv6 address gets deleted when interface goes down."
Problem
The kernel removes the IPv6 address if the linux interface is set to admin
"down". Upon forcing the interface to admin "up" state, the IPv6 addresses
are hence not found to be configured on the interface. The IPv6 addresses
should be present on the interface if the interface is forced to be admin
"up".

Root cause for the bug
We need to reprogram the configured IPv6 addresses once the interface is
forced admin "up". This should fix the current behavior.

Fix Details
To know when the kernel interfaces transition between "up" and "down",
we need to subscribe to the interface tables. Once we get an interface
"up"/"down" transition notification, we reprogram the kernel with the
IPv6 addresses on the interface (in case admin "up" is received. We do
this step only for the interfaces having L3 configuration only.

Misc
The code review was sent earlier as well in https://review.openhalon.io/#/c/1142/.
But this failed the merge attempt due to change in the file name l3portd_ipcfg.c
to portd_l3.c. I am abondoning the code review at https://review.openhalon.io/#/c/1142/.

Change-Id: I3f1146aea24cdddf1321c210a88f0f600b9a121d
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hp.com>