chg: dev: Add VRF handling code to portd
    Currently portd listens to swns name alone and with VRF addition
It has to listen to other namespaces also. The support to handle
other VRFs is being added in two phases. This changes are Phase1,
containing raw apis to configure interfaces in different namespaces
with a VRF_ENABLE flag which is disabled by default. Phase 2 would
handle removing the flag and handle link state notifications from
DB.

Change-Id: I08a56d14a89c35499bfa82b589750a00a42c1afd
Signed-off-by: Prerana B.H <prerana.bh@hpe.com>