fix: dev: [Moduler framework testcase]writing a retry function for few TCs
TG-1180
Fixing tc6 of test_portd_ct_functionality.py which is ipv6 address wan't configured in the kernel.
It takes some time for kernel to get configured with the ipv6 address so retrying in this kind of situations.
This is the cherry-pick from the rel/dill ref: https://review.openswitch.net/#/c/9461/


Change-Id: I51122e996b887e7b9a35b81ea2bc07fff08d22b9
Signed-off-by: Dinesh MC <dinesh.m-c@hpe.com>