fix: dev: Loopback fix to log the events properly.
This commit includes changes where to configure IP address for
a loopback interface the function portd_reconfig_ipaddr is called now.
This does the functionality of portd_reconfig_loopback_ipaddr which
was previously used and additionally also updates the loacl port
structure to have the ipv4 and ipv6 addresses. This updation was
missing previously and hence the event logs where not getting updated
correctly. This is required for supportability.

portd_reconfig_ns_loopback at one place was redundant, so removed
that call one place.

Also added two test cases for the Loopback devices- To delete IP and
verify in kernel and to delete IP when part of VRf and verify in the
kernel.

Change-Id: I0a2896fb355ba20c62e0f691c3a151215905d6b5
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>